Separate FilterControl from ResourceList, don't export types, export as _SECRET_INTERNAL component
