+++ src/components/FilterControl/FilterControl.scss
@@ -1,5 +1,5 @@
-@import '../../../../styles/common';
-@import '../../variables';
+@import '../../styles/common';
+@import 'variables';
 
 $spacing: tight;
