+++ src/components/ResourceList/components/index.ts
@@ -1,5 +1,3 @@
 export * from './BulkActions';
 
 export * from './CheckableButton';
-
-export * from './FilterControl';
